ScreenPosition and particle explosions
Add new struct ScreenPosition for screen
positions to work in the same fashion as grid positions
with the GridPosition. Move conversion function to the
struct by implementing From trait for it.

Add a simple particle system for the particle explosions.
When a bomb hasn't been eaten, and it explodes after a
timeout, show a simple particle explosion.

The bomb doesn't have any effect on the snake yet.